latest

+++ auth-server/src/main/java/com/ninePixels/poetry/authserver/userDetails/UserRepository.java
@@ -1,7 +1,7 @@
 package com.ninePixels.poetry.authserver.userDetails;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	 
     User findByUsername(String username);
 }
